pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = [
    "api",
    "util",
]
omit = [
    "tests/*",
    "venv/*",
    "*/migrations/*",
    "manage.py",
    "*/settings/*",
    "*/static/*",
    "*/templates/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[project]
name = "spotify-github-profile"
version = "0.1.0"
dependencies = [
    "flask==2.3.3",
    "werkzeug==2.3.8",
    "firebase-admin==6.5.0",
    "requests==2.32.3",
    "pillow>=10.0.0",
    "colorgram.py==1.2.0",
    "python-dotenv==1.0.0",
]
requires-python = ">=3.11"